+++ async_channel/channels/__init__.py
@@ -1,5 +1,4 @@
-# cython: language_level=3
-#  Drakkar-Software channel
+#  Drakkar-Software Async-Channel
 #  Copyright (c) Drakkar-Software, All rights reserved.
 #
 #  This library is free software; you can redistribute it and/or
@@ -14,18 +13,36 @@
 #
 #  You should have received a copy of the GNU Lesser General Public
 #  License along with this library.
+"""
+Define async_channel implementation and usage
+"""
+from async_channel.channels import channel_instances
+from async_channel.channels import channel
 
-from channel.channels cimport channel
-from channel.channels.channel cimport (
-    Channel
+from async_channel.channels.channel_instances import (
+    ChannelInstances,
+    set_chan_at_id,
+    get_channels,
+    del_channel_container,
+    get_chan_at_id,
+    del_chan_at_id,
 )
-
-from channel.channels cimport channel_instances
-from channel.channels.channel_instances cimport (
-    ChannelInstances
+from async_channel.channels.channel import (
+    Channel,
+    set_chan,
+    del_chan,
+    get_chan,
 )
 
 __all__ = [
     "ChannelInstances",
+    "set_chan_at_id",
+    "get_channels",
+    "del_channel_container",
+    "get_chan_at_id",
+    "del_chan_at_id",
     "Channel",
+    "set_chan",
+    "del_chan",
+    "get_chan",
 ]