+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.0.1] - 2020-10-01
 ### Added
 - Logging dynamic implementation
+- Project name to 'Async-Channel'
 
 ## [2.0.0] - 2020-10-01
 ### Update